+++ colliders.py
@@ -1,11 +1,12 @@
 import pygame
 
-from Jazz.utils import direction_to, dist_to, line_intersection, line_circle, Vec2
+from Jazz.baseObject import GameObject
+from Jazz.utils import Vec2, direction_to, dist_to, line_circle, line_intersection
 
 
-class Collider:
-    def __init__(self, pos):
-        self._pos = Vec2(pos)
+class Collider(GameObject):
+    def __init__(self, pos, **kwargs):
+        super().__init__(pos, "collider", **kwargs)
         if not hasattr(self, "vertices"):
             self.vertices = [Vec2(0, 0)]
         self.edges = []
@@ -47,16 +48,14 @@ class Collider:
                         new = False
                         break
                 if new:
-                    self.normals.append(
-                        Vec2(edge[1] - edge[0]).normalize().rotate(90)
-                    )
+                    self.normals.append(Vec2(edge[1] - edge[0]).normalize().rotate(90))
         else:
             self._left = -self.radius
             self._right = self.radius
             self._top = -self.radius
             self._bottom = self.radius
 
-    def draw(self, surface, offset=None):
+    def _draw(self, surface, offset=None):
         draw_verts = [self.pos + vert + offset for vert in self.vertices]
         pygame.draw.polygon(surface, self.color, draw_verts)
 
@@ -118,9 +117,7 @@ class Collider:
                         new = False
                         break
                 if new:
-                    self.normals.append(
-                        Vec2(edge[1] - edge[0]).normalize().rotate(90)
-                    )
+                    self.normals.append(Vec2(edge[1] - edge[0]).normalize().rotate(90))
 
     def rotate_around(self, degrees, center):
         center = Vec2(center)
@@ -207,15 +204,6 @@ class Collider:
             return depth, normal
         else:
             return depth, -normal
-           
-    
-    @property
-    def pos(self):
-        return self._pos
-
-    @pos.setter
-    def pos(self, new_pos):
-        self._pos = Vec2(new_pos)
 
     @property
     def top(self):
@@ -301,7 +289,12 @@ class CircleCollider(Collider):
         pygame.draw.circle(surface, self.color, self._pos + offset, self.radius, 1)
         pygame.draw.circle(surface, "red", self.pos + offset, 5)
         pygame.draw.circle(surface, "gray", self.pos + self._center + offset, 2)
-        pygame.draw.rect(surface, "yellow", (self.left + offset.x, self.top + offset.y, self.size[0], self.size[1]), 1)
+        pygame.draw.rect(
+            surface,
+            "yellow",
+            (self.left + offset.x, self.top + offset.y, self.size[0], self.size[1]),
+            1,
+        )
 
 
 class PolyCollider(Collider):
@@ -321,23 +314,24 @@ class PolyCollider(Collider):
             for i in range(self._size):
                 j = (i + 1) % self._size
                 self.edges.append([self.vertices[i], self.vertices[j]])
-                
+
+
 class RayCollider(Collider):
     def __init__(self, pos, length=1):
         self.vertices = [Vec2(0, 0), Vec2(length, 0)]
         self._length = length
         Collider.__init__(self, pos)
-    
+
     @property
     def length(self):
         return self._length
-    
+
     @length.setter
     def length(self, length):
         self._length = length
         self.vertices[1] = direction_to(self.vertices[0], self.facing) * length
         self.edges[0] = (self.vertices[0], self.vertices[1])
-    
+
     def collide_ray(self, collider):
         col_type = type(collider)
         if not isinstance(collider, Collider):
@@ -345,20 +339,21 @@ class RayCollider(Collider):
         if isinstance(collider, pygame.Rect):
             collider = RectCollider(collider.center, collider.width, collider.height)
         if collider is None:
-            raise TypeError(f'{col_type} is an invalid collider')
-        
+            raise TypeError(f"{col_type} is an invalid collider")
+
         if isinstance(collider, CircleCollider):
-            return line_circle(self.pos,
-                               self.pos + self.vertices[1],
-                               collider.pos,
-                               collider.radius)
+            return line_circle(
+                self.pos, self.pos + self.vertices[1], collider.pos, collider.radius
+            )
         else:
             collisions = []
             for edge in collider.edges:
-                point = line_intersection(self.pos,
-                                          self.pos + self.vertices[1],
-                                          collider.pos + edge[0],
-                                          collider.pos + edge[1])
+                point = line_intersection(
+                    self.pos,
+                    self.pos + self.vertices[1],
+                    collider.pos + edge[0],
+                    collider.pos + edge[1],
+                )
                 if point is not None:
                     collisions.append(point)
             if collisions:
@@ -366,9 +361,7 @@ class RayCollider(Collider):
                 closest_collision = Vec2()
                 for point in collisions:
                     if dist_to(self.pos, point) < closest_dist:
-                       closest_collision = point
-                       closest_dist = dist_to(self.pos, point)
+                        closest_collision = point
+                        closest_dist = dist_to(self.pos, point)
                 return closest_collision
         return None
-        
-
